requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
flask_login